% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_simulation_calculateSurvey.R
\name{calculateSurvey}
\alias{calculateSurvey}
\alias{calculateSurvey,FLBiol,FLFisheries,FLIndex-method}
\alias{calculateSurvey,FLBiol,FLFisheries,FLIndices-method}
\alias{calculateSurvey,FLBiols,FLFisheries,list-method}
\title{Calculate survey indices from \code{FLStocks} and \code{FLIndices}}
\usage{
calculateSurvey(stk, flt, idx, use_q = TRUE, use_time = TRUE, use_fastF = TRUE)

\S4method{calculateSurvey}{FLBiol,FLFisheries,FLIndex}(stk, flt, idx, use_q = TRUE, use_time = TRUE, use_fastF = TRUE)

\S4method{calculateSurvey}{FLBiol,FLFisheries,FLIndices}(stk, flt, idx, use_q = TRUE, use_time = TRUE, use_fastF = TRUE)

\S4method{calculateSurvey}{FLBiols,FLFisheries,list}(stk, flt, idx, use_q = TRUE, use_time = TRUE, use_fastF = TRUE)
}
\arguments{
\item{stk}{Object of class \code{FLBiol} or \code{FLBiols} containing
information from the operating model for one or more stocks.}

\item{flt}{Object of class \code{FLFisheries} containing information from the
operating model for one or more fishing fleets.}

\item{idx}{List, \code{FLIndices} or \code{FLIndex} object containing
survey information.}

\item{use_q}{Logical. Calculate survey indices by multipling stock numbers
at age with survey catchability? Defaults to \code{TRUE}.}

\item{use_time}{Logical. Use survey timing to correct available stock numbers
for fishing and natural mortality. Defaults to \code{TRUE}.}

\item{use_fastF}{Logical. Use fast c++ function to calculate fishing mortality
at age? Defaults to \code{TRUE}.}
}
\value{
an object of class \code{FLIndices} with the survey index stored in
the \code{index} slot.
}
\description{
Function calculate survey indices from the numbers-at-age contained in the
Operating model (\code{FLBiol}).
}
\details{
This is adapted from code used in WKNSMSE (2019).
}
