% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_effortdyn_FLBiol_v2.R
\name{catchBaranov}
\alias{catchBaranov}
\title{Calculation of fleet-stock total weight catches or landings given fleet efforts}
\usage{
catchBaranov(par, dat, adviceType, islog = FALSE)
}
\arguments{
\item{par}{Numeric vector. A vector of fleet effort on a log-scale or natural-scale}

\item{dat}{List. Input data containing numbers-at-age and natural mortality-at-age
for each stock, and landed and discarded mean weights-at-age,
fraction-landed-at-age, catch-selection-at-age, catchability-at-age
and quota-share for each combination of stock and fleet.}

\item{adviceType}{Character. Values may be 'catch' or 'landings'. Determines
whether values returned by the function are based on catch
or landings.}

\item{islog}{(Optional) Boolean. Is supplied effort (\code{par}) on a log-scale?
Defaults to \code{FALSE}.}
}
\value{
A matrix of the total weight of catch or landings with
fleets on columns and stocks on rows
}
\description{
This function calculates the total weight of catches or landings per stock
for each fleet given the inputted effort.
The function requires input data to be structured following the
format given by \code{FLBiols2List}.
}
\details{
Input data should contain the following variables listed within each iteration:

\itemize{
\item \code{n}: a named list
\itemize{
\item vector of numbers-at-age for ith stock
}
\item \code{m}: a named list
\itemize{
\item vector of natural mortality-at-age for ith stock
}
\item \code{landwt}: a named list
\itemize{
\item matrix of landed individual mean weight-at-age for ith stock (age x fleet)
}
\item \code{discwt}: a named list
\itemize{
\item matrix of discarded individual mean weight-at-age for ith stock (age x fleet)
}
\item \code{landfrac}: a named list
\itemize{
\item matrix of proportion of size-selective-retained catch-at-age for
ith stock (age x fleet)
}
\item \code{catchsel}: a named list
\itemize{
\item matrix of catch selectivity-at-age for ith stock (age x fleet)
}
\item \code{catchq}: a matrix of catchability (stock x fleet)
\item \code{quotashare}: a matrix of proportional quota-share (stock x fleet)
}
}
