% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_quality_control.R
\name{checkRun}
\alias{checkRun}
\title{Check whether the correct choke effort has been identified for each fleet}
\usage{
checkRun(object, yr, effortType = "min", sig.fig = 3, verbose = TRUE)
}
\arguments{
\item{object}{List. Complete outputs from a \code{MixME} simulations, comprising
an operating model (\code{om}), management procedure (\code{ctrl_obj}),
and \code{tracking}}

\item{yr}{Numeric. For which simulation year should the quality check be
performed?}

\item{effortType}{Char. 'min' or 'max'}

\item{sig.fig}{Integer. Residuals below this threshold are ignored.}

\item{verbose}{Should summaries be printed to console? Defaults to \code{TRUE}}
}
\value{
A dataframe containing year (\code{yr}), basis for calculating uptake performance,
(\code{metric}), fleet name (\code{fleet}), residual quota undershoot or overshoot that
exceed threshold (\code{residual}).
}
\description{
This function checks whether there is residual under- or over-shoot of quota
in a 'min' or 'max' simulation and returns a list of the fleets for optimised
effort does not correspond with expected quota uptake. If residuals fall
within tolerance limits, representing correct uptake, \code{TRUE} is returned.
}
\details{
In a 'min' simulation, residual quota for at least one stock should be zero
and an undershoot of all other stocks. Conversely, a 'max'
scenario should result in zero residual quota for at least one stock and
and overshoot of all other stocks.
}
