% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_effortdyn_FLBiol_v2.R
\name{effortBaranov}
\alias{effortBaranov}
\title{Calculation of fleet efforts given mixed fisheries technical interactions}
\usage{
effortBaranov(
  omList,
  adviceType = "catch",
  par = NULL,
  method = "nlminb",
  correctResid = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{omList}{List. Input data following the structure returned by \code{FLBiols2List}
containing numbers-at-age and natural mortality-at-age
for each stock, and landed and discarded mean weights-at-age,
fraction-landed-at-age, catch-selection-at-age, catchability-at-age
and quota-share for each combination of stock and fleet.}

\item{adviceType}{Character. Values may be 'catch' or 'landings'. Determines
whether inputted advice values correspond to catch or
landings. Defaults to 'catch'.}

\item{par}{(Optional) Numeric vector. A vector of fleet effort on a log-scale.}

\item{method}{(Optional) Character. Determines the optimisation algorithm that is used.
Values may be 'nlminb' or 'Nelder-Mead'. Default is 'nlminb'.
HIGHLY recommended that the default is used.}

\item{correctResid}{(Optional) Boolean. Should fleet efforts be scaled down
to bring residual overquota catch to zero if residual
over-quota catch following optimisation exceeds 0.001?
Defaults to \code{FALSE}}

\item{parallel}{(Optional) Boolean. Should function be run in parallel?}
}
\value{
A list containing outputs for the final optimisation.
}
\description{
This function calculates the effort for each fleet required to catch their
first quota.
}
\details{
Outputs are a list of the form:

\itemize{
\item A list of iterations
\itemize{
\item \code{par}: A vector of log-scale effort per fleet
\item \code{objective}: The final optimisation objective function value.
\item \code{convergence}: The final optimisation convergence code value.
0 indicates successful convergence.
\item \code{iterations}: The final optimisation number of iterations.
\item \code{message}: The final optimisation convergence message.
}
}
}
