% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{EM.mix}
\alias{EM.mix}
\title{EM Algorithm for Gaussian Mixture Model Combined with Linear Model}
\usage{
EM.mix(
  X,
  y,
  lambda,
  muK,
  sigK,
  beta,
  sigma2,
  tol = 1e-05,
  maxit = 5000,
  verb = FALSE,
  is.profile = FALSE
)
}
\arguments{
\item{X}{Surrogate data or observations in main study}

\item{y}{Responses in outcome model}

\item{lambda}{A vector of size K containing initial value of mixing probabilities}

\item{muK}{A matrix of size K * p containing initial values of component mean, K-means center is specified if NULL}

\item{sigK}{An array with size p*p*K containing K p*p matrix, each of which is initial values of component variance,
K-means variance is specified if NULL}

\item{beta}{A vector size K containing initial linear coefficients}

\item{sigma2}{A scalar of initial variance of random noise in linear model}

\item{tol}{The convergence criterion}

\item{maxit}{The maximum number of iterations}

\item{verb}{If true, then various updates are printed during each iteration of the algorithm}

\item{is.profile}{If true, estimate nuisance parameters based on profile likelihood method}
}
\value{
A list with the following elements:
\itemize{
\item{pi} The final mixing probabilities
\item{mu} The final mean vectors
\item{sigK} The final variance matrix of each component
\item{beta} The final linear coefficients
\item{sigma2} The final variance of random noise
\item{posterior} Membership probability
\item{iter} Number of iterations
}
}
\description{
EM Algorithm for Gaussian Mixture Model Combined with Linear Model
}
