% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{EM.mixme}
\alias{EM.mixme}
\title{EM Algorithm for Gaussian Mixture Model Combined with Measurement Error Model and Linear Model}
\usage{
EM.mixme(
  Zm,
  Xv,
  Zv,
  y,
  lambda,
  muK,
  sigK,
  alpha,
  A,
  sigE,
  beta,
  sigma2,
  tol = 1e-05,
  maxit = 5000,
  verb = FALSE,
  is.profile = FALSE
)
}
\arguments{
\item{Zm}{Surrogate data in main study}

\item{Xv}{True observations in validation study}

\item{Zv}{Surrogate data in validation study}

\item{y}{Responses in outcome model}

\item{lambda}{A vector of size K containing initial value of mixting probabilities}

\item{muK}{A matrix of size K * p containing initial values of component mean, K-means center is specified if NULL}

\item{sigK}{An array with size p*p*K containing K p*p matrix, each of which is initial values of component variance,
K-means variance is specified if NULL}

\item{alpha}{A vector of size p consisting of intercept in measurement error model}

\item{A}{A matrix of size p*p consisting of slope in measurement error model}

\item{sigE}{A matrix of size p*P consisting of variance of noise in measurement error model}

\item{beta}{A vector size K containing initial linear coefficients}

\item{sigma2}{A scalar of initial variance of random noise in linear model}

\item{tol}{The convergence criterion}

\item{maxit}{The maximum number of iterations}

\item{verb}{If true, then various updates are printed during each iteration of the algorithm}

\item{is.profile}{Used for computing profile variance. If true, beta and sigma2 will not be estimated}

\item{profile}{A vector containing profile parameters}
}
\value{
A list with the following elements:
\itemize{
\item{pi} The final mixing probabilities
\item{mu} The final mean vectors
\item{sigK} The final variance matrix of each component
\item{alpha} The final \code{alpha} in measurement error model
\item{A} The final \code{A} in measurement error model
\item{sigK} The final \code{sigK} in measurement error model
\item{beta} The final linear coefficients
\item{sigma2} The final variance of random noise
\item{posterior} Membership probability
\item{iter} Number of iterations
}
}
\description{
EM Algorithm for Gaussian Mixture Model Combined with Measurement Error Model and Linear Model
}
