% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_estForecast.R
\name{estForecast}
\alias{estForecast}
\title{Short-term forecast to estimate stock in advice year}
\usage{
estForecast(
  stk,
  ctrl,
  args,
  forecast,
  fwd_trgt,
  fwd_yrs,
  fwd_yrs_average,
  fwd_yrs_rec_start,
  fwd_yrs_sel,
  fwd_yrs_lf_remove,
  fwd_splitLD
)
}
\arguments{
\item{stk}{an object of class \code{FLBiol} or \code{FLStock}}

\item{ctrl}{a forward control object of class \code{...}
for each stock.}

\item{args}{list of additional arguments}

\item{tracking}{a named list of tracking objects to monitor emergent dynamic
properties. Must contain an \code{FLStock} object named 'stk'
containing landings, discards, catch and harvest information.}
}
\description{
Function runs a short-term forecast for a single stock to estimate stock
properties in the advice year.
}
