% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_est.R
\name{estMixME}
\alias{estMixME}
\title{Implementation of stock estimation methods}
\usage{
estMixME(
  stk,
  idx,
  ctrl = NULL,
  om = NULL,
  args,
  estmethod = NULL,
  tracking,
  fitList = NULL,
  fwdList = NULL
)
}
\arguments{
\item{stk}{a list of stocks, each of class \code{FLBiol} or \code{FLStock},
or a single stock to be assessed}

\item{idx}{a list of indices ...}

\item{ctrl}{a forward control object of class \code{...}
for each stock.}

\item{args}{list of additional arguments}

\item{tracking}{a named list of tracking objects to monitor emergent dynamic
properties. Must contain an \code{FLStock} object named 'stk'
containing landings, discards, catch and harvest information.}

\item{fitList}{(optional) a named list of length \emph{n} stocks. Each list element is a sub-list
containing the arguments needed to estimate stock status.
Defaults to \code{NULL}.}

\item{fwdList}{(optional) a named list of length \emph{n} stocks. Each list element is a sub-list
containing the arguments needed to carry out a short-term
forecast for the stock. See 'Details' for required list elements.
Defaults to \code{NULL}.}

\item{estMethod}{named list of estimation methods for each stock.
List elements may be the name of an in-built estimation
method or a user-supplied function that is applied to the
stock. If \code{NULL}, defaults to 'perfectObs'.}
}
\description{
Function takes a time-series of observed catch or index data
(\code{FLStock} and \code{FLIndex} respectively) and estimates current stock
status using user-supplied methods. Optionally, if a lag exists between the
assessment and management years, a short-term forecast may be carried out to
estimate stock status during the management year.
}
\details{
The \emph{fwdList} argument is a list that should contain the following elements:
\itemize{
\item \emph{forecast}: Logical. Should a forecast be carried out?
\item \emph{fwd_trgt}: Character. The short-term forecast target type. Either
'fsq' or 'TAC'.
\item \emph{fwd_yrs}: Integer. The number of years to forecast
\item \emph{fwd_yrs_average}: Integer vector. Index of historic years over which to
average biological data
\item \emph{fwd_yrs_rec_start}: Integer vector. Index of historic years from which
sample recruitment.
\item \emph{fwd_yrs_sel}: Integer vector. Index of historic years over which to average
selectivity-at-age
\item \emph{fwd_yrs_lf_remove}: Integer vector. Index of historic years for which
landings fraction data are overwritten by the landings
fraction data in the most recent year.
\item \emph{fwd_splitLD}: Separate landings and discards numbers in forecast outputs.
}
}
