% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_fitSRMixME.R
\name{fitSRMixME}
\alias{fitSRMixME}
\title{Wrapper function to fit stock-recruit models}
\usage{
fitSRMixME(
  om,
  rec_yrs_start = NULL,
  rec_model = "segreg",
  iy,
  ny,
  uncertainty = FALSE,
  verbose = TRUE
)
}
\description{
Wrapper function to fit stock-recruit models
}
