% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_effortdyn_FLBiol_v2.R
\name{FLBiols2List}
\alias{FLBiols2List}
\title{Convert object into simple nested list structure}
\usage{
FLBiols2List(om, year, advice, useCpp = TRUE)
}
\arguments{
\item{om}{Operating model}

\item{year}{Integer. Year in which advice is implemented.}

\item{adviceType}{Character. Takes either 'landings' or 'catch'. If 'landings',
quota-limits apply to landed weight. If 'catch', quota-limits
apply to overall catch weight (comprising both landings and discards).

NOTE! CURRENTLY ASSUMES THAT CATCH.Q IS FLPAR... MIGHT WANT TO CHANGE THIS...
NOTE! R version CURRENTLY ASSUMES THAT ALL FLEETS CATCH ALL STOCKS ... FIX THIS!}
}
\description{
This function converts the linked \code{FLBiols} and \code{FLFisheries}
objects into a set of nested lists with the following structure:
}
\details{
\itemize{
\item \code{iterations}: a named list
\itemize{
\item \code{n}: a named list
\itemize{
\item vector of numbers-at-age for ith stock
}
\item \code{m}: a named list
\itemize{
\item vector of natural mortality-at-age for ith stock
}
\item \code{landwt}: a named list
\itemize{
\item matrix of landed individual mean weight-at-age for ith stock (age x fleet)
}
\item \code{discwt}: a named list
\itemize{
\item matrix of discarded individual mean weight-at-age for ith stock (age x fleet)
}
\item \code{landfrac}: a named list
\itemize{
\item matrix of proportion of size-selective-retained catch-at-age for
ith stock (age x fleet)
}
\item \code{catchsel}: a named list
\itemize{
\item matrix of catch selectivity-at-age for ith stock (age x fleet)
}
\item \code{catchq}: a matrix of catchability (stock x fleet)
\item \code{quotashare}: a matrix of proportional quota-share (stock x fleet)
}
}
}
