% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_FLFleet2FLFishery.R
\name{FLFleet2FLFishery}
\alias{FLFleet2FLFishery}
\title{Convert \code{FLFleets} class object into \code{FLFisheries} class object}
\usage{
FLFleet2FLFishery(fleets, na_replace = 0, verbose = TRUE)
}
\arguments{
\item{fleets}{An object of class \code{FLFleets}}

\item{na_replace}{Value used to replace instances of \code{NA}. Defaults to
0.}

\item{verbose}{Should progress be printed to screen? Defaults to \code{TRUE}.}
}
\value{
object of class \code{FLFisheries}
}
\description{
Function converts \code{FLFleets} objects with a nested \code{FLMetiers} and
\code{FLCatches} hierarchy into \code{FLFisheries} objects with nested
\code{FLCatches} by averaging across metiers.
}
