% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_fwd.R
\name{fwd_MIME}
\alias{fwd_MIME}
\title{Forward projection of stocks - including process error}
\usage{
fwd_MIME(
  stks,
  flts,
  ctrl,
  sr,
  sr.residuals,
  sr.residuals.mult = TRUE,
  maxF = 2,
  proc_res = NULL,
  ...
)
}
\arguments{
\item{stks}{Object of class \code{FLBiols}}

\item{flts}{Object of class \code{FLFisheries}}

\item{ctrl}{Control object.}
}
\value{
A list containing
}
\description{
This function takes the output of the implementation error model and
projects the stocks and fleets forward one time step.
}
