% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_fwd.R
\name{fwdMixME}
\alias{fwdMixME}
\title{Forward projection of stocks - including process error}
\usage{
fwdMixME(
  om,
  args,
  tracking,
  sr_residuals = NULL,
  proc_res = NULL,
  adviceType,
  effortType,
  exceptions,
  multiplier,
  nyear = NULL,
  effort_max = 100,
  effsharemethod = NULL,
  effshareargs = NULL,
  ...
)
}
\arguments{
\item{om}{A list of \code{FLBiols} and \code{FLFisheries} containing the relevant
stock and fleet information in the Operating Model.}

\item{args}{list of additional arguments}

\item{tracking}{a named list of tracking objects to monitor emergent dynamic
properties. Contains catch advice for each stock.}

\item{sr_residuals}{(Optional) \code{FLQuants} or \code{list}. Multiplicative
stock recruitment residuals. Defaults to \code{NULL}.}

\item{proc_res}{(Optional) Character. Where is process error noise stored?
If \code{NULL}, no process error is applied to stock numbers.
Defaults to \code{NULL}.}

\item{adviceType}{Character. The basis of management advice.
Can be 'catch' or 'landings'.'f' is not yet possible.}

\item{effortType}{Character. The basis of effort constraint. Can be 'max', 'min'
or 'sqE'.}

\item{exceptions}{Matrix (stock, fleet). Can be 0 or 1.}

\item{multiplier}{Matrix (stock, fleet). Defaults to 1.}

\item{nyear}{(Optional) Number of reference years for the calculation of
status quo effort. Defaults to \code{NULL}.}

\item{effort_max}{(Optional) Integer indicating the maximum allowed
fishing effort for any fleet. Defaults to 100 and
cannot be \code{NULL}.}

\item{effsharemethod}{(Optional) User-supplied function to carry out redistribution
of effort-share across metiers. Must take \code{om} and
\code{yr} as minimum inputs.}

\item{effshareargs}{(Optional) List of additional arguments required by effort-share
redistribution function.}
}
\value{
A list containing the \code{FLBiols} and \code{FLFisheries} objects
with projected stock numbers, fleet efforts, and fleet-stock landings
and discards numbers.
}
\description{
This function takes the output of the implementation error model and
projects the stocks and fleets forward one time step.
}
