% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_simulation_getFage.R
\name{getFage}
\alias{getFage}
\title{Calculate fishing mortality at age for a defined stock}
\usage{
getFage(stks, flts, stkname, yr = NULL, use_fastF = TRUE)
}
\arguments{
\item{stks}{Object of class \code{FLBiols}.}

\item{flts}{Object of class \code{FLFisheries}.}

\item{stkname}{Character. Name of the stock for which fishing mortality at
age is required}

\item{yr}{(Optional) Integer vector. Vector of years for which to subset the
calculated fishing mortality at age.}

\item{use_fastF}{(Optional). Use C++ implementation of fishing mortality
calculation? Defaults to \code{TRUE}.}
}
\value{
A 6-dimensional array containing the fishing mortality at age for a user-defined
stock.
}
\description{
This function calculates fishing mortality at age given an \code{FLFisheries}
object.
}
