% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gmm}
\alias{gmm}
\title{Gaussian Mixture Model}
\usage{
gmm(X, K, lambda, muK, sigK, tol = 1e-05, maxit = 5000, verb = TRUE)
}
\arguments{
\item{X}{A matrix consisting of the data}

\item{K}{Number of components}

\item{lambda}{A vector of size K containing initial value of mixting probabilities}

\item{muK}{A matrix of size K * p containing initial values of component mean, K-means center is specified if NULL}

\item{sigK}{An array with size p*p*K containing K p*p matrix, each of which is initial values of component variance,
K-means variance is specified if NULL}

\item{tol}{The convergence criterion}

\item{maxit}{The maximum number of iterations}

\item{verb}{If true, then various updates are printed during each iteration of the algorithm}
}
\value{
A list with the following elements:
\itemize{
\item{pi} The final mixing probabilities
\item{mu} The final mean vectors
\item{sigK} The final variance matrix of each component
\item{posterior} Membership probability
\item{iter} Number of iterations
}
}
\description{
Simultaneously estimate parameters in Gaussian mixture model combined with linear model
}
