% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_hcr.R
\name{hcrFixedC}
\alias{hcrFixedC}
\title{Fixed Catch advice implementation}
\usage{
hcrFixedC(stk, Ctrg, args)
}
\arguments{
\item{stk}{Observed stocks. An object of class \code{FLStock}.}

\item{Ctrg}{Fixed catch target. An object of class \code{FLQuant} or numeric
value.}

\item{args}{Additional arguments.}
}
\value{
A control object containing the adjusted F target for the stock.
}
\description{
Generates a control object to implement fixed Catch advice.
}
\author{
Adapted from WKNSMSE (2019) code by S. H. Fischer.
}
