% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_hcrRun.R
\name{hcrRun}
\alias{hcrRun}
\title{Wrapper for advice rule methods}
\usage{
hcrRun(
  stk,
  flt,
  idx,
  args,
  hcrpars = NULL,
  hcrmethod = NULL,
  ctrg = NULL,
  ftrg = NULL,
  tracking
)
}
\arguments{
\item{stk}{list of stocks}

\item{hcrpars}{(optional) Harvest Control Rule parameters for each stock. A named
nested list. Each named list element corresponds to a stock
and should contain a list of parameters for that stock.
Defaults to \code{NULL}.}

\item{hcrmethod}{named list of advice rule methods for each stock.
List elements may be the name of an in-built advice rule
method or a user-supplied function that is applied to the
stock. Defaults to 'hcrICES'.}

\item{ctrg}{(optional) named list of target catch for each stock.}

\item{ftrg}{(optional) named list of target fishing mortality for each stock}

\item{tracking}{a named list of tracking objects to monitor emergent dynamic
properties.}
}
\description{
...
}
