% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_isys.R
\name{isysForecast}
\alias{isysForecast}
\title{Short-term forecast to translate F-based advice to TAC}
\usage{
isysForecast(
  stk,
  sr,
  ctrl,
  args,
  tracking,
  forecast = NULL,
  fwd_trgt = NULL,
  fwd_yrs = NULL,
  fwd_yrs_average = NULL,
  fwd_yrs_rec_start = NULL,
  fwd_yrs_sel = NULL,
  fwd_yrs_lf_remove = NULL,
  fwd_splitLD = NULL
)
}
\arguments{
\item{stk}{an object of class \code{FLBiol} or \code{FLStock}}

\item{ctrl}{a forward control object of class \code{...}
for each stock.}

\item{args}{list of additional arguments}

\item{tracking}{a named list of tracking objects to monitor emergent dynamic
properties. Must contain an \code{FLStock} object named 'stk'
containing landings, discards, catch and harvest information.}

\item{forecast}{Logical. Should a forecast be carried out?}

\item{fwd_trgt}{Character. The short-term forecast target type. Either
'fsq' or 'TAC'.}

\item{fwd_yrs}{Integer. The number of years to forecast.}

\item{fwd_yrs_average}{Integer vector. Index of historic years over which to
average biological data.}

\item{fwd_yrs_rec_start}{Integer vector. Index of historic years from which
sample recruitment.}

\item{fwd_yrs_sel}{Integer vector. Index of historic years over which to average
selectivity-at-age.}

\item{fwd_yrs_lf_remove}{Integer vector. Index of historic years for which
landings fraction data are overwritten by the landings
fraction data in the most recent year.}

\item{fwd_splitLD}{Separate landings and discards numbers in forecast outputs.}
}
\description{
Function runs a short-term forecast for a single stock to translate a target
fishing mortality into a catch target.
}
\details{
If input stock object is of class \code{FLBiol}, function first coerces to
\code{FLStock}. Catch and harvesting properties are taken from either the
operating model or observation error model tracking object. A forecast is
carried out using \code{fwd} and the catch target for the advice year is
extracted.
}
