% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_isys.R
\name{isysMixME}
\alias{isysMixME}
\title{Implementation of advice outputs}
\usage{
isysMixME(
  stk,
  sr = NULL,
  ctrl,
  args,
  isysmethod = NULL,
  tracking,
  fwdList = NULL
)
}
\arguments{
\item{stk}{object of class \code{FLBiols} containing a named \code{FLBiol}
for each stock.}

\item{ctrl}{named list containing a forward control object of class \code{...}
for each stock.}

\item{args}{list of additional arguments}

\item{isysmethod}{named list of implementation system methods for each stock.
List elements may be the name of an in-built method or a
user-supplied function that is applied to the
stock. Defaults to class \code{NULL}.}

\item{tracking}{a named list of tracking objects to monitor emergent dynamic
properties}

\item{fwdList}{a named list of stocks. Each list element is a sub-list
containing the arguments needed to carry out a short-term
forecast for the stock. See 'Details'.}
}
\value{
a named list containing a list of forward control objects for each stock
and a list of tracking objects.
}
\description{
Function takes harvest control rule outputs (in the form of a control object)
and applies the implementation system to generate the advice that will drive
fleet activity. The implementation system can be any set of formal methods
that modify advice (e.g. TAC change limits, banking and borrowing) and/or
transform advice targets from catch-based to effort-based (f-based) or vice
versa.
}
\details{
The \emph{fwdList} argument is a list that should contain the following elements:
\itemize{
\item \emph{forecast}: Logical. Should a forecast be carried out?
\item \emph{fwd_trgt}: Character. The short-term forecast target type. Either
'fsq' or 'TAC'.
\item \emph{fwd_yrs}: Integer. The number of years to forecast
\item \emph{fwd_yrs_average}: Integer vector. Index of historic years over which to
average biological data
\item \emph{fwd_yrs_rec_start}: Integer vector. Index of historic years from which
sample recruitment.
\item \emph{fwd_yrs_sel}: Integer vector. Index of historic years over which to average
selectivity-at-age
\item \emph{fwd_yrs_lf_remove}: Integer vector. Index of historic years for which
landings fraction data are overwritten by the landings
fraction data in the most recent year.
\item \emph{fwd_splitLD}: Separate landings and discards numbers in forecast outputs.
}
}
