% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_fwd.R
\name{makeFCB}
\alias{makeFCB}
\title{Automatic generation of FCB matrix for FLasher}
\usage{
makeFCB(biols, flts)
}
\description{
This function takes an FLBiols and FLFisheries object and generates an FCB
matrix
}
