% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_multiSAM2FLFishery.R
\name{makeFLCatch}
\alias{makeFLCatch}
\title{\subsection{Extract and process landings and discards numbers and weights, selectivity, and fishing mortality}{
}}
\usage{
makeFLCatch(
  SAMfit,
  qnt,
  useSAMcatch,
  yearRange,
  variates,
  res_n,
  land_disc_idx,
  catch_idx,
  fishing_idx
)
}
\description{
This function extracts the relevant catch information for a given fleet from
a fitted multifleet SAM model. This is called once per fleet within 'multiSAM2FLFishery'
}
\details{
The outputted FLCatch object may be used to define a new FLFishery and
FLFisheries, define a new FLFishery within an existing FLFisheries, or be
appended to an existing FLFishery. Therefore the function may be called within
any of these three contexts for a given stock as multiSAM2FLFishery loops over
each fleet.

Supplying the correct indices to the function is critical to appropriately
subsetting the data for the correct fleet.
}
