% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_makeMixME.R
\name{makeMixME}
\alias{makeMixME}
\title{Function to assemble inputs to a MixME simulation. Outputs a list with correct
arguments for the specified simulation. Object slots may require updating.}
\usage{
makeMixME(
  om,
  catch_obs = NULL,
  index_obs = NULL,
  catch_dev = NULL,
  index_dev = NULL,
  management_lag = 1,
  management_type = "HCR",
  effort_type = "min",
  parallel = FALSE
)
}
\description{
Function to assemble inputs to a MixME simulation. Outputs a list with correct
arguments for the specified simulation. Object slots may require updating.
}
