% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mereg.lm}
\alias{mereg.lm}
\title{Regression Calibration Model}
\usage{
mereg.lm(Zm, Xv, Zv, y, K, tol = 1e-05, maxit = 5000, verb = TRUE)
}
\arguments{
\item{Zm}{Surogate data in the main study}

\item{Xv}{True observations in the validation study}

\item{Zv}{Surogate data in the validation study}

\item{y}{A vector consisting of outcome of linear model}

\item{K}{number of clusters}

\item{tol}{The convergence criterion}

\item{maxit}{The maximum number of iterations}

\item{verb}{If true, then various updates are printed during each iteration of the algorithm}
}
\value{
output of linear outcome model
}
\description{
Regression Calibration Model
}
