% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixme.R
\name{mereg}
\alias{mereg}
\title{Regression Calibration Model}
\usage{
mereg(Zm, Xv, Zv, lambda, muK, sigK, tol = 0.001, maxit = 5000, verb = FALSE)
}
\arguments{
\item{Zm}{Surogate data in the main study}

\item{Xv}{True observations in the validation study}

\item{Zv}{Surogate data in the validation study}

\item{lambda}{A vector of size K containing initial value of mixting probabilities}

\item{muK}{A matrix of size K * p containing initial values of component mean, K-means center is specified if NULL}

\item{sigK}{An array with size p*p*K containing K p*p matrix, each of which is initial values of component variance,
K-means variance is specified if NULL}

\item{tol}{The convergence criterion}

\item{maxit}{The maximum number of iterations}

\item{verb}{If true, then various updates are printed during each iteration of the algorithm}
}
\value{
output of mixtools::mvnormalmixEM
}
\description{
Regression Calibration Model
}
