% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mix.init}
\alias{mix.init}
\title{Initialization for Various EM Algorithms}
\usage{
mix.init(X = NULL, K = NULL, Xv = NULL, Zv = NULL)
}
\arguments{
\item{X}{Data in main study}

\item{K}{Number of clusters}

\item{Xv}{Observations in validation study}

\item{Zv}{Surrogate data in validation study}
}
\value{
A list with the following elements:
\itemize{
\item{pi} Initial estimate of mixing probabilities 
\item{mu} A K*p matrix consisting of initial estimate of Gaussian mean 
\item{sigK} An K*p*p array consisting of initial estimate of Gaussian variance
\item{alpha} Initial estimate of intercept in measurement error model
\item{A} Initial estimate of slope in measurement error model
\item{sigE} Initial estimate of variance in measurement error model
}
}
\description{
Initialization for Various EM Algorithms
}
