% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixme.R
\name{mix.variance}
\alias{mix.variance}
\title{Variance-covariance Estimate by Profile Likelihood Method}
\usage{
mix.variance(X, y, lambda, muK, sigK, beta, sigma2)
}
\arguments{
\item{X}{A matrix consisting of the data}

\item{y}{A vector consisting of outcome of linear model}

\item{lambda}{A vector of size K containing initial value of mixting probabilities}

\item{muK}{A matrix of size K * p containing initial values of component mean, K-means center is specified if NULL}

\item{sigK}{An array with size p*p*K containing K p*p matrix, each of which is initial values of component variance,
K-means variance is specified if NULL}

\item{beta}{A vector size K containing initial linear coefficients}

\item{sigma2}{A scalar of initial variance of random noise in linear model}
}
\value{
A variance-covariance matrix of (\code{beta}, \code{sigma2})
}
\description{
The variance of \code{beta} and \code{sigma} is of interest, others are taken as nuisance parameters
}
