% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixME-package.R
\name{MixME-package}
\alias{MixME-package}
\alias{MixME}
\title{MixME (Mixed-fishery Multi-stock Evaluation)}
\description{
An R package for mixed fisheries Management Strategy Evaluation (MSE) using multi-stock and multi-fleet Operating Models.
}
\details{
MixME:
\itemize{
\item can carry out full MSE to evaluate the performance of Management Procedures.
\item can carry out smaller-scale simulations with perfect stock observations to evaluate advice rule behaviour.
can accommodate a range of fleet and stock parameter uncertainties.
\item makes use of FLR libraries (\url{https://github.com/flr})
\item has built-in helper functions to condition Operating Models from fitted SAM
(State-space Assessment Model; \url{https://github.com/fishfollower/SAM})
models with uncertainty
\item has built-in diagnostic tools to check for errors occuring in the Management
Procedure (e.g. assessment model fitting errors) or the Operating Model
(e.g. fleet effort optimisation)
}
}
