% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixme.R
\name{mixme.lm}
\alias{mixme.lm}
\title{Gaussian Mixture Model Combined with Measurement Error and Linear model}
\usage{
mixme.lm(
  Zm,
  Xv,
  Zv,
  y,
  K,
  lambda,
  muK,
  sigK,
  alpha,
  A,
  sigE,
  beta,
  sigma2,
  tol = 0.001,
  maxit = 5000,
  verb = TRUE,
  variance = TRUE
)
}
\arguments{
\item{Zm}{Surogate data in the main study}

\item{Xv}{True observations in the validation study}

\item{Zv}{Surogate data in the validation study}

\item{y}{A vector consisting of outcome of linear model}

\item{K}{Number of components}

\item{lambda}{A vector of size K containing initial value of mixting probabilities}

\item{muK}{A matrix of size K * p containing initial values of component mean, K-means center is specified if NULL}

\item{sigK}{An array with size p*p*K containing K p*p matrix, each of which is initial values of component variance,
K-means variance is specified if NULL}

\item{alpha}{A vector of size p consisting of intercept in measurement error model}

\item{A}{A matrix of size p*p consisting of slope in measurement error model}

\item{sigE}{A matrix of size p*P consisting of variance of noise in measurement error model}

\item{beta}{A vector size K containing initial linear coefficients}

\item{sigma2}{A scalar of initial variance of random noise in linear model}

\item{tol}{The convergence criterion}

\item{maxit}{The maximum number of iterations}

\item{verb}{If true, then various updates are printed during each iteration of the algorithm}

\item{variance}{If true, variance of \code{beta} and \code{sigma2} are estimated based on profile likelihood method}
}
\value{
A list with the following elements:
\itemize{
\item{pi} The final mixing probabilities
\item{mu} The final mean vectors
\item{sigK} The final variance matrix of each component
\item{alpha} The final \code{alpha} in measurement error model
\item{A} The final \code{A} in measurement error model
\item{sigK} The final \code{sigK} in measurement error model
\item{beta} The final linear coefficients
\item{sigma2} The final variance of random noise
\item{posterior} Membership probability
\item{iter} Number of iterations
}
}
\description{
simultaneously estimate all the parameters
}
