% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_profile.R
\name{mixme.variance}
\alias{mixme.variance}
\title{Variance-covariance Estimate by Profile Likelihood Method}
\usage{
mixme.variance(
  Zm,
  Xv,
  Zv,
  y,
  lambda,
  muK,
  sigK,
  alpha,
  A,
  sigE,
  beta,
  sigma2,
  maxit
)
}
\arguments{
\item{Zm}{Surogate data in the main study}

\item{Xv}{True observations in the validation study}

\item{Zv}{Surogate data in the validation study}

\item{y}{A vector consisting of outcome of linear model}

\item{lambda}{A vector of size K containing initial value of mixting probabilities}

\item{muK}{A matrix of size K * p containing initial values of component mean, K-means center is specified if NULL}

\item{sigK}{An array with size p*p*K containing K p*p matrix, each of which is initial values of component variance,
K-means variance is specified if NULL}

\item{alpha}{A vector of size p consisting of intercept in measurement error model}

\item{A}{A matrix of size p*p consisting of slope in measurement error model}

\item{sigE}{A matrix of size p*P consisting of variance of noise in measurement error model}

\item{beta}{A vector size K containing initial linear coefficients}

\item{sigma2}{A scalar of initial variance of random noise in linear model}

\item{maxit}{The maximum number of iterations}
}
\value{
A variance-covariance matrix of (\code{beta}, \code{sigma2})
}
\description{
The variance of \code{beta} and \code{sigma} is of interest, others are taken as nuisance parameters
}
