% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_MixME2FLStocks.R
\name{MixME2FLStocks}
\alias{MixME2FLStocks}
\title{Convert \code{MixME} operating model into \code{FLFisheries} class object}
\usage{
MixME2FLStocks(om, units_harvest = "f")
}
\arguments{
\item{om}{A MixME om comprising a named lists of \code{stks} and \code{flts} with the
classes \code{FLBiols} and \code{FLFisheries}, respectively.}
}
\value{
object of class \code{FLStocks}
}
\description{
Function converts a \code{MixME} operating model containing \code{stks} (\code{FLBiols})
and \code{flts} (\code{FLFisheries}) into \code{FLStocks} by aggregating fleet data.
}
\examples{
## load example data
data("mixedfishery_MixME_input")

## collapse Operating Model to FLStocks
res <- MixME2FLStocks(mixedfishery_MixME_input$om)
}
