% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_multiSAM2FLFishery.R
\name{multiSAM2FLFishery}
\alias{multiSAM2FLFishery}
\title{Convert a SAM fitted object into an \code{FLFisheries} object}
\usage{
multiSAM2FLFishery(
  SAMfit,
  stkname = NULL,
  useSAMcatch = TRUE,
  add = FALSE,
  fleets = NULL,
  yearRange = NULL
)
}
\arguments{
\item{SAMfit}{SAM fitted stock assessment model object of class \code{sam}}

\item{stkname}{Character. Stock name associated with SAM object.}

\item{useSAMcatch}{Optional argument. If \code{TRUE}, the fitted catches estimated
by SAM are used. Otherwise, the observed catches are used.
Defaults to \code{TRUE}.}

\item{add}{Optional argument. Defaults to \code{FALSE}. If \code{TRUE},
\code{FLCatch} information is appended onto an existing \code{FLFishery}
structure}

\item{fleets}{Optional argument. An object of class \code{FLFisheries} to which
fitted SAM object data are extracted and appended.}

\item{yearRange}{Optional argument to extend the FLStock year dimension to
a user-supplied range. Consists of an integer vector
of two elements. First element is the minimum year. Second
element is the maximum year.}
}
\value{
An \code{FLFishery} object
}
\description{
This function takes a SAM fitted stock assessment object as input and
returns a \code{FLFisheries} object.
}
\details{
Returned \code{FLFisheries} object contains data for one or more
\code{FLFishery} objects. Each \code{FLFishery} object contains the following
slots:
\itemize{
\item capacity (\code{catch.q})
\item effort (\code{landings.n})
\item landings mean weight-at-age (\code{landings.wt})
\item proportion catch retained-at-age (\code{landings.sel})
\item discards numbers-at-age (\code{discards.n})
\item discards mean weight-at-age (\code{discards.wt})}

Additionally, each \code{FLFishery} object contains one or more \code{FLCatch}
object with the following slots:

\itemize{
\item landings numbers-at-age (\code{landings.n})
\item landings mean weight-at-age (\code{landings.wt})
\item discards numbers-at-age (\code{discards.n})
\item discards mean weight-at-age (\code{discards.wt})
\item proportional selectivity-at-age (\code{catch.sel})
\item mean price-per-unit-weight-at-age (\code{price})}

NOTE: The partial fishing mortality-at-age for each stock is inserted in
the respective catchability-at-age slot. Further processing into
catchability-at-age is needed.
}
\section{Warning}{

This function requires \code{FLCore} to operate.
}

