% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_multiSAM2FLFleet.R
\name{multiSAM2FLFleet}
\alias{multiSAM2FLFleet}
\title{Convert a SAM fitted object into an \code{FLFleets} object}
\usage{
multiSAM2FLFleet(
  SAMfit,
  stkname = NULL,
  useSAMcatch = TRUE,
  add = FALSE,
  fleets = NULL,
  yearRange = NULL
)
}
\arguments{
\item{SAMfit}{SAM fitted stock assessment model object of class \code{sam}}

\item{stkname}{Character. stock name associated with SAM object.}

\item{useSAMcatch}{Optional argument. If \code{TRUE}, the fitted catches estimated
by SAM are used. Otherwise, the observed catches are used.
Defaults to \code{TRUE}.}

\item{add}{Optional argument. Defaults to \code{FALSE}. If \code{TRUE},
\code{FLCatch} information is appended onto an existing \code{FLFleet}
and \code{FLMetier} structure}

\item{fleets}{Optional argument. An object of class \code{FLFleets} to which
fitted SAM object data are extracted and appended.}

\item{yearRange}{Optional argument to extend the FLStock year dimension to
a user-supplied range. Consists of an integer vector
of two elements. First element is the minimum year. Second
element is the maximum year.}
}
\value{
An \code{FLFleets} object
}
\description{
This function takes a SAM fitted stock assessment object as input and
returns a \code{FLFleets} object.
}
\details{
Returned \code{FLFleet} object contains data for:
\itemize{
\item partial fishing mortality-at-age (\code{catch.q})
\item landings numbers-at-age (\code{landings.n})
\item landings mean weight-at-age (\code{landings.wt})
\item proportion catch retained-at-age (\code{landings.sel})
\item discards numbers-at-age (\code{discards.n})
\item discards mean weight-at-age (\code{discards.wt})}

Fbar range is ignored because it does not provide meaningful information
at the partial fishing mortality level.

NOTE: The partial fishing mortality-at-age for each stock is inserted in
the respective catchability-at-age slot. Further processing into
catchability-at-age is needed.
}
\section{Warning}{

This function requires \code{FLCore} and \code{FLFleets} to operate.
}

