% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_multiSAM2FLIndex.R
\name{multiSAM2FLIndex}
\alias{multiSAM2FLIndex}
\title{Extract survey data from a SAM fitted object to an \code{FLIndices} object}
\usage{
multiSAM2FLIndex(
  SAMfit,
  yearRange = NULL,
  uncertainty = FALSE,
  niter = 1000,
  samVariates = NULL,
  seed = NULL
)
}
\arguments{
\item{SAMfit}{A fitted SAM model of class \code{sam}.}

\item{yearRange}{Defaults to \code{NULL}.}

\item{uncertainty}{\code{Logical}. Should replicates be stochastically sampled from
the joint likelihood? Defaults to \code{FALSE}.}

\item{niter}{\code{Integer}. The number of replicates to be sampled if \code{uncertainty}
is \code{TRUE}. Defaults to 1000.}

\item{samVariates}{(Optional) Pre-sampled replicates. If supplied, these values
are used instead of re-sampling from the joint likelihood.
This is useful for generating \code{FLStock}, \code{FLFishery} and
\code{FLIndex} objects with consistent replicate values.}

\item{seed}{(Optional) \code{numeric}. A random seed for reproducibility when
sampling replicates.}
}
\value{
an \code{FLIndices} object
}
\description{
Function takes a fitted SAM object as input and returns an \code{FLIndices}
object containing the observed survey index values and the fitted
stock catchability.
}
\examples{
## download stock assessment data
library(stockassessment)
had <- fitfromweb("HAD7bk_2020_Assessment")

## Extract indices (without uncertainty)
hadIdx <- multiSAM2FLIndex(had, uncertainty = FALSE)

\donttest{
## Extract indices (with uncertainty)
hadIdx <- multiSAM2FLIndex(had, uncertainty = TRUE, niter = 1000)
}
}
