% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_multiSAM2FLStock.R
\name{multiSAM2FLStock}
\alias{multiSAM2FLStock}
\title{Convert a SAM fitted object into a \code{FLStock} object}
\usage{
multiSAM2FLStock(
  SAMfit,
  useSAMcatch = TRUE,
  yearRange = NULL,
  uncertainty = FALSE,
  niter = 1000,
  samVariates = NULL,
  seed = NULL
)
}
\arguments{
\item{SAMfit}{A SAM fitted stock assessment model object}

\item{useSAMcatch}{Optional argument. If \code{TRUE}, the fitted catches estimated
by SAM are used. Otherwise, the observed catches are used.
Defaults to \code{TRUE}.}

\item{yearRange}{Optional argument to extend the FLStock year dimension to
a user-supplied range. Consists of an integer vector
of two elements. First element is the minimum year. Second
element is the maximum year.}

\item{uncertainty}{Optional argument. If \code{TRUE}, the iteration dimension
of the returned objects contains replicates (uncertainty)
sampled from the variance-covariance matrix of the fitted
SAM object. Defaults to \code{FALSE}}

\item{niter}{Optional integer. The number of replicates sampled if
\code{uncertainty} is \code{TRUE}. Default is 1000 replicates.}

\item{seed}{Optional integer. This is a random number seed to generate
reproducible outputs. Defaults to \code{NULL}.}
}
\value{
An \code{FLStock} object
}
\description{
This function takes a SAM fitted stock assessment object as input and
returns an \code{FLStock} object.
}
\details{
Returned \code{FLStock} object contains data for:
\itemize{
\item stock numbers-at-age (\code{stock.n})
\item catch numbers-at-age (\code{catch.n})
\item landings numbers-at-age (\code{landings.n})
\item discards numbers-at-age (\code{discards.n})
\item proportion mature-at-age (\code{mat})
\item natural mortality-at-age (\code{m})
\item fishing mortality-at-age (\code{harvest})
\item stock mean weight-at-age (\code{stock.wt})
\item catch mean weight-at-age (\code{catch.wt})
\item landings mean weight-at-age (\code{landings.wt})
\item discards mean weight-at-age (\code{discards.wt})}

Catch numbers and fishing mortality-at-age are summed across commercial fleets,
whereas catch, landings and discards mean weight-at-age are averages weighted
by the catch number proportions by commercial fleets. Landings and discards
numbers-at-age are calculated using the catch proportion weighted mean
landings fraction.

The function also updates the Fbar range for the stock from the fitted
SAM object.

If \code{uncertainty} is \code{TRUE}, the first replicate of the iteration
dimension contains the best-fit SAM estimates for stock numbers, harvest rate
and catch, whereas the
}
\section{Warning}{

This function requires \code{FLCore} to operate.
}

