% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_multiSAM2FLR.R
\name{multiSAMfay}
\alias{multiSAMfay}
\title{\subsection{Extract fleet F-at-age from sampled SAM parameters}{
}}
\usage{
multiSAMfay(SAMfit, variates, niter, catch_fleets = NULL)
}
\description{
Function takes a fitted SAM object and previously sampled parameter variates
and coerces the fleet F-at-age into an array
}
