% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_multiSAM2FLR.R
\name{multiSAMvariates}
\alias{multiSAMvariates}
\title{\subsection{Sample variates from fitted SAM object}{
}}
\usage{
multiSAMvariates(SAMfit, niter)
}
\description{
Function takes a fitted SAM object and samples a number of variates for each
estimated parameter from a multivariate normal distribution. Function returns
a matrix.
}
