% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_oem.R
\name{oemMixME}
\alias{oemMixME}
\title{Implementation of stock observation methods}
\usage{
oemMixME(
  om,
  deviances,
  observations,
  args,
  tracking,
  catch_timing = NULL,
  idx_timing = NULL,
  use_stk_oem = FALSE,
  use_catch_residuals = FALSE,
  use_idx_residuals = FALSE
)
}
\arguments{
\item{observations}{}
}
\description{
Function ...
}
\details{
Note that this function is intended to be looped over, and therefore is run
separately for each stock.
}
