% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_oemRun.R
\name{oemRun}
\alias{oemRun}
\title{Wrapper function for stock observation methods}
\usage{
oemRun(
  om,
  deviances,
  observations,
  args,
  tracking,
  catch_timing = NULL,
  idx_timing = NULL,
  use_stk_oem = FALSE,
  use_catch_residuals = FALSE,
  use_idx_residuals = FALSE
)
}
\arguments{
\item{om}{The operating model representing the true state and dynamics of
the system. A named list comprising stock information (\code{FLStocks})
and fishing fleet information (\code{FLFisheries}).}

\item{deviances}{Environmental noise in stock catch and survey data for the
historical and projection periods. A nested named list
with the elements \code{stk} and \code{idx}. See details.}

\item{observations}{The structure of stock catch and survey data used to
estimate stock status.
A nested named list with the elements \code{stk}
and \code{idx}.}

\item{args}{List of additional arguments}

\item{tracking}{A named list of tracking objects to monitor emergent dynamic
properties}

\item{catch_timing}{A named list of catch timing relative to assessment year}

\item{idx_timing}{A named list of survey index timing relative to assessment year}

\item{use_stk_oem}{Logical vector. Length \code{n} stocks. Should a supplied
Observation Error Model structure be used?}

\item{use_catch_residuals}{Logical vector. Length \code{n} stocks. Should
catch residuals be used to generate uncertainty?}

\item{use_idx_residuals}{Logical vector. Length \code{n} stocks. Should
survey index residuals be used to generate uncertainty?}
}
\value{
A named list of stock observations, survey indices and updated tracking
object
}
\description{
Function calls stock observation method for each stock/management plan.
}
