% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_oemTrim.R
\name{oemTrimFLBiol}
\alias{oemTrimFLBiol}
\title{Trim FLBiol and FLFisheries to data period}
\usage{
oemTrimFLBiol(stk0, flt0, x, ay, idx_timing, catch_timing)
}
\arguments{
\item{stk0}{An observed \code{FLBiol}}

\item{flt0}{An observed \code{FLFisheries}}

\item{ay}{Numeric. The assessment year.}

\item{idx_timing}{A named list of survey index timing relative to assessment year}

\item{catch_timing}{A named list of catch timing relative to assessment year}
}
\value{
A named list of stock observations.
}
\description{
Function takes stock observations in the form of \code{FLBiol} and \code{FLFisheries}
and trims the year dimension to the period for which we have catch or survey data.
}
