% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_oemTrim.R
\name{oemTrimFLFisheries}
\alias{oemTrimFLFisheries}
\title{Trim FLFisheries to data period}
\usage{
oemTrimFLFisheries(object, minyr, maxyr)
}
\arguments{
\item{object}{An observed \code{FLFisheries}}

\item{maxyr}{Numeric. The final data year.}
}
\value{
A truncated \code{FLFisheries} object.
}
\description{
Function takes fisheries observations in the form of \code{FLFisheries}
and trims the year dimension to the period for which we have data.
}
