% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_oemTrim.R
\name{oemTrimFLStock}
\alias{oemTrimFLStock}
\title{Trim FLStock to data period}
\usage{
oemTrimFLStock(stk0, x, ay, idx_timing, catch_timing)
}
\arguments{
\item{stk0}{An observed \code{FLStock}}

\item{ay}{Numeric. The assessment year.}

\item{idx_timing}{A named list of survey index timing relative to assessment year}

\item{catch_timing}{A named list of catch timing relative to assessment year}
}
\value{
A named list of stock observations.
}
\description{
Function takes stock observations in the form of \code{FLStock} and trims
the year dimension to the period for which we have catch or survey data.
}
