% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_hcr.R
\name{phcrMixME}
\alias{phcrMixME}
\title{Initialise parameters for the ICES Harvest Control Rule}
\usage{
phcrMixME(stk, args, hcrpars, tracking)
}
\arguments{
\item{args}{named list of stocks. Each list element contains a named list
of Harvest Control Rule parameters.}
}
\value{
a named list of stocks. Each list element contains an \code{FLPar}
object for stock Harvest Control Rule parameters
}
\description{
Function takes a named list of Harvest Control Rule parameters for each stock
and coerces parameters into \code{FLPar} objects to be used by the Harvest Control
Rule.
}
