% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot_catchabilities.R
\name{plot_catchability}
\alias{plot_catchability}
\title{Catchability conditioning plot}
\usage{
plot_catchability(
  FLFisheries,
  fisheries = "all",
  stocks = "all",
  years = "all",
  outlier_minmax = c(0.001, 0.999)
)
}
\arguments{
\item{FLFisheries}{is an \code{object} of class FLFisheries}

\item{fisheries}{is a \code{character} vector of either 'all' or the name of the fisheries to include in the figure.}

\item{stocks}{is a \code{character} vector of either 'all' or the name of the stocks to include in the figure.}

\item{years}{is a \code{numeric} or \code{character} vector with the years to include (can be 'all')}

\item{outlier_minmax}{is a \code{numeric} vector of length 2, of the minimum and maximum quantiles for catchability by stock that are used to identify outliers.}
}
\value{
a \code{plot} with catchability by year for each stock, and fishery (colours) and area (shapes) combinations.
The outliers are identified as the red points, extreme values lie in the red shaded area. AND a \code{dataframe} with the
identified outliers and extreme values.
}
\description{
\code{plot_catchability} Function takes an FLFisheries object and returns a plot of the catchabilities by stock,
including identifying any outliers and extreme values. This can be useful to check conditioning before running the model.
}
\examples{
\donttest{
data("mixedfishery_MixME_input")
fleets <- mixedfishery_MixME_input$om$flts
plot_catchability(FLFisheries = fleets, fleets = 'all', stocks = 'all', outlier_minmax = c(0.001, 0.999))
}

}
