% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot_catch.R
\name{plot_catches}
\alias{plot_catches}
\title{Total Catch conditioning plot}
\usage{
plot_catches(
  FLFisheries,
  fisheries = "all",
  stocks = "all",
  years = "all",
  type = "combined",
  values = "weight"
)
}
\arguments{
\item{FLFisheries}{is an \code{object} of class FLFisheries}

\item{fisheries}{is a \code{character} vector of either 'all' or the name of the fisheries to include in the figure.}

\item{stocks}{is a \code{character} vector of either 'all' or the name of the stocks to include in the figure.}

\item{type}{is a \code{character} vector of either \code{combined} to aggregate across fisheries, or \code{individual} to plot by fishery.}

\item{values}{is a \code{character} vector of either \code{numbers} to plot catch numbers-at-age or \code{weight}}

\item{`years`}{is a \code{numeric} or \code{character} vector with the years to include (can be 'all).}
}
\value{
a plot with total catch (tonnage) by year for each stock, which can be faceted by individual fisheries.
Landings and discards are denoted by different fill colours in the bars.
}
\description{
\code{plot_catches} Function takes an FLFisheries object and returns a plot of the total catches by stock or by stock and fishery.
This can be useful to check conditioning before running the model.
}
\examples{
\donttest{
data("mixedfishery_MixME_input")
fleets <- mixedfishery_MixME_input$om$flts
plot_catches(FLFisheries = fleets, fisheries = "all", stock = "all", years = "all", type = "individual")
}

}
