% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot_discard_rates.R
\name{plot_discard_rates}
\alias{plot_discard_rates}
\title{Discard rate conditioning plot}
\usage{
plot_discard_rates(
  FLFisheries,
  fisheries = "all",
  stocks = "all",
  years = "all",
  type = "combined"
)
}
\arguments{
\item{FLFisheries}{is an \code{object} of class FLFisheries}

\item{fisheries}{is a \code{character} vector of either 'all' or the name of the fisheries to include in the figure.}

\item{stocks}{is a \code{character} vector of either 'all' or the name of the stocks to include in the figure.}

\item{years}{is a \code{numeric} or \code{character} vetor of the years to include (can be 'all').}

\item{type}{is a \code{character} vector of either \code{combined} to aggregate across fisheries, or \code{individual} to plot by fishery.}
}
\value{
a plot with discard rate (fraction discarded) by year for each stock, which can be faceted by individual fisheries.
areas within a fishery are denoted by different linetypes and shapeds.
}
\description{
\code{plot_discard_rates} Function takes an FLFisheries object and returns a plot of the discard rate by stock or by stock and fishery.
This can be useful to check conditioning before running the model.
}
\examples{
\donttest{
data("mixedfishery_MixME_input")
fleets <- mixedfishery_MixME_input$om$flts
plot_discard_rates(FLFisheries = fleets, fisheries = "all", stock = "all", type = "individual")
}

}
