% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot_effort.R
\name{plot_effort}
\alias{plot_effort}
\title{Effort conditioning plot}
\usage{
plot_effort(FLFisheries, fisheries = "all", years = "all", type = "effort")
}
\arguments{
\item{FLFisheries}{is an \code{object} of class FLFisheries}

\item{fisheries}{is a \code{character} vector of either 'all' or the name of the fisheries to include in the figure.}

\item{type}{is a \code{character} vector of either \code{effort} to aggregate across area, or \code{shares} to plot fleet-area effort shares by fishery.}

\item{`years`}{is a \code{numeric} or \code{character} vector with the years to include (can be 'all).}
}
\value{
a plot with either effort by year for each area or effort shares.
}
\description{
\code{plot_effort} Function takes an FLFisheries object and returns a plot of the effort by area and fishery.
This can be useful to check conditioning before running the model.
}
\examples{
\donttest{
data("mixedfishery_MixME_input")
fleets <- mixedfishery_MixME_input$om$flts
plot_effort(FLFisheries = fleets, fisheries = "all", type = "effort")
}

}
