% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot_quotashare.R
\name{plot_quotashare}
\alias{plot_quotashare}
\title{quota share conditioning plot}
\usage{
plot_quotashare(FLFisheries, fisheries = "all", stocks = "all", years = "all")
}
\arguments{
\item{FLFisheries}{is an \code{object} of class FLFisheries}

\item{fisheries}{is a \code{character} vector of either 'all' or the name of the fisheries to include in the figure.}

\item{stocks}{is a \code{chacter} vector of either 'all' or the name of the stocks to include in the figure}

\item{`years`}{is a \code{numeric} or \code{character} vector with the years to include (can be 'all).}
}
\value{
a plot with either effort by year for each area or effort shares.
}
\description{
\code{plot_quotashare} Function takes an FLFisheries object and returns a plot of the quota share by fishery and stock.
This can be useful to check conditioning before running the model.
}
\examples{
\donttest{
data("mixedfishery_MixME_input")
fleets <- mixedfishery_MixME_input$om$flts
plot_quotashare(FLFisheries = fleets, fisheries = "all", stocks = "all", years = "all")
}

}
