% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot_selectivity.R
\name{plot_selectivity}
\alias{plot_selectivity}
\title{Selectivity conditioning plot}
\usage{
plot_selectivity(
  FLFisheries,
  fisheries = "all",
  years = "all",
  stocks = "all",
  type = "combined"
)
}
\arguments{
\item{FLFisheries}{is an \code{object} of class FLFisheries}

\item{fisheries}{is a \code{character} vector of either 'all' or the name of the fisheries to include in the figure.}

\item{years}{is a \code{numeric} or \code{character} vector with the years to include (can be 'all')}

\item{stocks}{is a \code{character} vector of either 'all' or the name of the stocks to include in the figure.}

\item{type}{is a \code{character} vector of either \code{individual} (by fleet and area) or \code{combined} (aggregated by stock).}
}
\value{
a plot with selectivity-at-age by year for each stock, or by individual fisheries, areas and years for each stock.
A smooth GAM is fit to indicate the selectivity as a function of age by stock and year or stock, fishery, area and year.
}
\description{
\code{plot_selectivity} Function takes an FLFisheries object and returns a plot of the seletcity-at-age by stock and year,
or by fishery, area and stock and year. This can be useful to check conditioning before running the model.
}
\examples{
\donttest{
data("mixedfishery_MixME_input")
fleets <- mixedfishery_MixME_input$om$flts
plot_selectivity(FLFisheries = fleets, fisheries = 'all', stocks = 'all', type = "combined")
}

}
