% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot_weights.R
\name{plot_weights}
\alias{plot_weights}
\title{Weight conditioning plot}
\usage{
plot_weights(
  FLFisheries,
  fisheries = "all",
  stocks = "all",
  years = "all",
  type = "catch.wt"
)
}
\arguments{
\item{FLFisheries}{is an \code{object} of class FLFisheries}

\item{fisheries}{is a \code{character} vector of either 'all' or the name of the fisheries to include in the figure.}

\item{stocks}{is a \code{character} vector of either 'all' or the name of the stocks to include in the figure.}

\item{years}{is a \code{numeric} or \code{character} vector of the years to include (can be 'all')}

\item{type}{is a \code{character} vector of either \code{landings.wt}, \code{discards.wt} or \code{catch.wt}.}
}
\value{
a plot with weight-at-age by year for each stock, with individual fisheries and areas indicated by the points.
A smooth GAM is fit to indicate the median weight as a function of age by stock and year.
}
\description{
\code{plot_weights} Function takes an FLFisheries object and returns a plot of the weights-at-age by stock.
This can include landings weight, discard weight and catch weight. This can be useful to check conditioning before running the model.
}
\examples{
\donttest{
data("mixedfishery_MixME_input")
fleets <- mixedfishery_MixME_input$om$flts
plot_weights(FLFisheries = fleets, fisheries = 'all', stocks = 'all', type = "catch.wt")
}

}
