% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixme.R
\name{prob.member}
\alias{prob.member}
\title{Compute Membership Probability for Given Observation}
\usage{
prob.member(x, lambda, muK, sigK)
}
\arguments{
\item{x}{A vector or a matrix, if x is a matrix, each row is taken to be a quantile}

\item{lambda}{A vector of size K consisting of mixing probabilities}

\item{muK}{A vector of Gaussian mean}

\item{sigK}{A matrix of Gaussain variance matrix}
}
\value{
A vector of size K consisting of membership probabilities
}
\description{
Compute Membership Probability for Given Observation
}
