% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_projection_Fastfwd.R
\name{projectFastfwd}
\alias{projectFastfwd}
\title{Project stock forward one year using a bespoke c++ function}
\usage{
projectFastfwd(
  om,
  tracking,
  pars,
  yr,
  popType = NULL,
  sr_residuals = NULL,
  proc_res = NULL
)
}
\arguments{
\item{om}{Operating model}

\item{tracking}{Tracking object}

\item{pars}{Numeric vector. Vector of optimised fleet efforts}

\item{yr}{Integer. year}
}
\value{
A named list containing the operating model and tracking object.
}
\description{
Function uses a bespoke c++ function to carry out a short-term forecast.
}
