% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_projection_FLasher.R
\name{projectFLasher}
\alias{projectFLasher}
\title{Project stock forward one year using FLasher}
\usage{
projectFLasher(
  om,
  tracking,
  pars,
  yr,
  effort_max,
  sr_residuals = NULL,
  proc_res = NULL,
  testfwd,
  adviceType
)
}
\arguments{
\item{om}{Operating model}

\item{tracking}{Tracking object}

\item{pars}{Numeric vector. Vector of optimised fleet efforts}

\item{yr}{Integer. year}

\item{effort_max}{Maximum allowed fleet effort}

\item{sr_residuals}{(Optional) Stock-Recuitment noise}

\item{proc_res}{(Optional) Process error}
}
\value{
A named list containing the operating model and tracking object.
}
\description{
Function uses the \code{FLasher} package to carry out a short-term forecast.
}
