% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMixME.R
\name{runMixME}
\alias{runMixME}
\title{Run mixed fishery management strategy simulation}
\usage{
runMixME(om, oem, ctrl_obj, args, ...)
}
\arguments{
\item{om}{operating model (OM)}

\item{oem}{observation error model (OEM)}

\item{args}{named list. Additional MSE simulation arguments.}

\item{mp}{management procedure (MP)}
}
\value{
A named list containing the projected operating model and a tracking
object containing summary metrics for observed, estimated and true
stock and fishery properties and simulation performance statistics.
}
\description{
This function carries out a single run for a set of one or more management
procedures (MP) to evaluate MP performance given mixed fishery technical
interactions. Function uses a conditioned operating model (OM) to simulate
true system dynamics, whereas the observation error model (OEM), MP and
additional arguments drive the imperfect evaluation of stock status and the
generation of management advice.
}
