% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_options_SAMassessment.R
\name{SAMassessment}
\alias{SAMassessment}
\title{Estimate stock status used the SAM state-space assessment model}
\usage{
SAMassessment(
  stk,
  idx,
  tracking,
  args,
  parallel = FALSE,
  conf = NULL,
  par_ini = NULL,
  track_ini = FALSE,
  forecast = FALSE,
  newtonsteps = 0,
  rel.tol = 0.001,
  fwd_trgt = "fsq",
  fwd_yrs = 1,
  fwd_yrs_average = -3:-1,
  fwd_yrs_rec_start = NULL,
  fwd_yrs_sel = -3:-1,
  fwd_yrs_lf_remove = -2:-1,
  fwd_splitLD = TRUE,
  ...
)
}
\arguments{
\item{stk}{Object of class \code{FLStock} containing observed stock
information including commercial catch data, individual mean
weights and biological parameters.}

\item{idx}{Object of class \code{FLIndices} containing survey indices}

\item{tracking}{Tracking object}

\item{args}{Additional arguments}

\item{parallel}{Logical. Should model fitting be parallelised? Defaults to
\code{FALSE}}

\item{conf}{List. SAM configuration}

\item{par_ini}{List. Initial parameter values for SAM}

\item{track_ini}{Logical. Should fitted parameters be stored for use as
initial values next year? Defaults to \code{FALSE}}

\item{forecast}{Logical. Should a short-term forecast be carried out? Defaults
to \code{TRUE}}

\item{fwd_trgt}{Character. Catch or effort target to use during forecast.
Defaults to 'fsq'.}

\item{fwd_yrs}{Integer. The number of years to forecast. Defaults to 1.}

\item{fwd_yrs_average}{Integer vector. The historical data years over which
biological parameter are averaged for use during
forecast. Defaults to -3:-1.}

\item{fwd_yrs_rec_start}{Integer. Starting historical year from which to sample
projection period recruitment during forecast.}

\item{fwd_yrs_sel}{Integer vector. The historical data years over which
catch selection-at-age is averaged for use during forecast.
Defaults to -3:-1.}

\item{fwd_yrs_lf_remove}{Integer Vector. ... Defaults to -2:-1.}

\item{fwd_splitLD}{Logical. Defaults to \code{TRUE}}
}
\value{
a named list containing the assessed stock and updated tracking
object.
}
\description{
Stock status estimation using bridging functions from the \code{FLfse} to
fit a SAM state-space model. Time-series of commercial catch and
survey indices are used as observation data to which the model is fitted.
}
