% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_options_SAMimplementation.R
\name{SAMimplementation}
\alias{SAMimplementation}
\title{Advice implementation using the SAM state-space assessment model}
\usage{
SAMimplementation(
  stk,
  tracking,
  ctrl,
  args,
  fwd_trgt = c("fsq", "hcr"),
  fwd_yrs = 2,
  fwd_yrs_average = -3:0,
  fwd_yrs_rec_start = NULL,
  fwd_yrs_sel = -3:-1,
  fwd_yrs_lf_remove = -2:-1,
  fwd_splitLD = TRUE
)
}
\arguments{
\item{stk}{Object of class \code{FLStock} containing observed stock
information including commercial catch data, individual mean
weights and biological parameters.}

\item{tracking}{Tracking object}

\item{ctrl}{A \code{FLasher} forward control object containing the outputs from the
Harvest Control Module.}

\item{args}{Additional arguments}

\item{fwd_trgt}{Character. Catch or effort target to use during forecast.
Defaults to 'fsq'.}

\item{fwd_yrs}{Integer. The number of years to forecast. Defaults to 1.}

\item{fwd_yrs_average}{Integer vector. The historical data years over which
biological parameter are averaged for use during
forecast. Defaults to -3:-1.}

\item{fwd_yrs_rec_start}{Integer. Starting historical year from which to sample
projection period recruitment during forecast.}

\item{fwd_yrs_sel}{Integer vector. The historical data years over which
catch selection-at-age is averaged for use during forecast.
Defaults to -3:-1.}

\item{fwd_yrs_lf_remove}{Integer Vector. ... Defaults to -2:-1.}

\item{fwd_splitLD}{Logical. Defaults to \code{TRUE}}

\item{forecast}{Logical. Should a short-term forecast be carried out? Defaults
to \code{TRUE}}
}
\value{
A named \code{list} containing the updated forward control object and the
updated tracking object for the stock.
}
\description{
Carry out a short-term forecast using the SAM state-space assessment model
to transform advised fishing mortality into an advised catch target.
}
