% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMixME.R
\name{simMixME}
\alias{simMixME}
\title{Implement simulation loop}
\usage{
simMixME(om, oem, tracking, ctrl_obj, args)
}
\arguments{
\item{om}{operating model (OM)}

\item{oem}{observation error model (OEM)}

\item{ctrl_obj}{management procedure (MP)}

\item{args}{named list. Additional MSE simulation arguments.}
}
\value{
A named list containing the projected operating model and a tracking
object containing summary metrics for observed, estimated and true
stock and fishery properties and simulation performance statistics.
}
\description{
This function carries out the MixME simulation loop, iterating over each
projection year. The function print progress through each year and management
procedure module, and returns an updated operating model and tracking object
}
