% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_extract_slot_fisheries.R
\name{slot_fisheries}
\alias{slot_fisheries}
\title{Extract FLFisheries}
\usage{
slot_fisheries(fisheries, slot.)
}
\arguments{
\item{fisheries}{An FLFisheries object with historical (and optionally conditioned) parameter values.}

\item{slot.}{\code{character} vector of length=1. The name of the slot to extract as a dataframe.}
}
\value{
is a dataframe with outputs from the slot for all FLFisheries in the object.
}
\description{
\code{slot_fisheries} Function takes an FLFisheries object and a named slot to extract the slot as a dataframe.
}
\examples{
\donttest{
data("mixedfishery_MixME_input")
fleets <- mixedfishery_MixME_input$om$flts
slot_fisheries(fisheries = fleets, slot. = "catch.q")
}

}
