% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_stfMixME.R
\name{stfQuant}
\alias{stfQuant}
\title{\subsection{Function to handle different methods of extending stock/fleet properties to future years}{
}}
\usage{
stfQuant(
  object,
  method,
  samples = NULL,
  ni = NULL,
  ignoreZero = FALSE,
  ignoreAllZero = FALSE
)
}
\description{
\subsection{Function to handle different methods of extending stock/fleet properties to future years}{
}
}
