% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_scenario_exceptionsMatrix.R
\name{stockExceptionsMatrix}
\alias{stockExceptionsMatrix}
\title{Generate scenario 'exceptions' matrix}
\usage{
stockExceptionsMatrix(om, yr, stockname = NULL)
}
\arguments{
\item{om}{List. A MixME operating model comprising two elements: 'stks' of
class \code{FLBiols} and 'flts' of class \code{FLFIsheries}.}

\item{yr}{Numeric. The year in the operating model that will be used to
determine if a given fleet has catchability for each stock.}

\item{stockname}{Char (Optional). The name of a single stock that is assumed
to limit the effort of all fleets that catch it. Useful for
setting-up stock-specific scenarios.}
}
\value{
a matrix with stocks on the rows and fleets on the columns.
}
\description{
The 'exceptions' matrix defines which stocks are effort-limiting for each
fleet. This simple function generate an exceptions matrix of the correct
dimensions for a given \code{MixME} operating model.
}
