% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_summary.R
\name{summary_catch_MixME}
\alias{summary_catch_MixME}
\title{Generate summaries of Operating Model fleet-stock catches}
\usage{
summary_catch_MixME(
  object,
  minyr = NULL,
  maxyr = NULL,
  fltnames = NULL,
  stknames = NULL
)
}
\description{
Function takes an Operating model as input and generate a summary dataframe
of fleet-stock catches for one or more fleets
}
