% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_summary.R
\name{summary_effort_MixME}
\alias{summary_effort_MixME}
\title{Generate summaries of Operating Model fleet efforts}
\usage{
summary_effort_MixME(object, minyr = NULL, maxyr = NULL, fltnames = NULL)
}
\description{
Function takes an Operating model as input and generate a summary dataframe
of effort for one or more fleets
}
