% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_summary.R
\name{summary_f_MixME}
\alias{summary_f_MixME}
\title{Generate summaries of Operating Model fishing mortality at age}
\usage{
summary_f_MixME(object, minyr = NULL, maxyr = NULL, stknames = NULL)
}
\description{
Function takes an Operating model as input and generate a summary dataframe
of mean fishing mortality at age for one or more stocks
}
