% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_summary.R
\name{summary_fbar_MixME}
\alias{summary_fbar_MixME}
\title{Generate summaries of Operating Model mean fishing mortality}
\usage{
summary_fbar_MixME(object, minyr = NULL, maxyr = NULL, stknames = NULL)
}
\description{
Function takes an Operating model as input and generate a summary dataframe
of mean fishing mortaltiy for one or more stocks
}
