% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_summary.R
\name{summary_risk_MixME}
\alias{summary_risk_MixME}
\title{Generate summaries of Operating Model risk}
\usage{
summary_risk_MixME(object, minyr = NULL, maxyr = NULL, stknames = NULL, Refpts)
}
\description{
Function takes an Operating model as input and generate a summary dataframe
of risk (the annual probability of SSB falling below Blim) for one or more stocks
}
