% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_summary.R
\name{summary_ssb_MixME}
\alias{summary_ssb_MixME}
\title{Generate summaries of Operating Model spawning stock biomass}
\usage{
summary_ssb_MixME(object, minyr = NULL, maxyr = NULL, stknames = NULL)
}
\description{
Function takes an Operating model as input and generate a summary dataframe
of spawning stock biomass for one or more stocks.
}
