% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_summary.R
\name{summary_uptake_MixME}
\alias{summary_uptake_MixME}
\title{Generate summaries of Operating Model stock quota uptake}
\usage{
summary_uptake_MixME(
  object,
  minyr = NULL,
  maxyr = NULL,
  fltnames = NULL,
  stknames = NULL
)
}
\description{
Function takes an Operating model as input and generate a summary dataframe
of quota uptake for one or more stocks
}
