% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_calculateQuotashare.R
\name{calculateQuotashare}
\alias{calculateQuotashare}
\title{Estimate stock quota-share from historic landings-share}
\usage{
calculateQuotashare(stks, flts, verbose = TRUE, useCpp = TRUE)
}
\description{
Estimate stock quota-share from historic landings-share
}
