% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_effortdyn_FLBiol_v2.R
\name{effortBaranov}
\alias{effortBaranov}
\title{Calculation of fleet efforts given mixed fisheries technical interactions}
\usage{
effortBaranov(
  omList,
  adviceType = "catch",
  effortType = "min",
  exceptions,
  multiplier,
  par = NULL,
  maxRetry = 1,
  useEffortAsInit = FALSE,
  useGlobalAsInit = FALSE,
  useTMB = TRUE,
  correctResid = FALSE,
  parallel = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{omList}{List. Input data following the structure returned by \code{FLBiols2List}
containing numbers-at-age and natural mortality-at-age
for each stock, and landed and discarded mean weights-at-age,
fraction-landed-at-age, catch-selection-at-age, catchability-at-age
and quota-share for each combination of stock and fleet.}

\item{adviceType}{Character. Values may be 'catch' or 'landings'. Determines
whether inputted advice values correspond to catch or
landings. Defaults to 'catch'.}

\item{effortType}{Character. Values may be 'min' or 'max'. Determines whether
effort optimisation should find the most-limiting (min) or the
least-limiting stock (max).}

\item{exceptions}{Numeric matrix (stocks, fleets). A matrix of effort-limiting
(1) and non-effort-limiting (0) stocks for each fleet}

\item{multiplier}{Numeric matrix (stocks, fleets). A matrix of quota-multipliers
for each combination of stock and fleet.}

\item{par}{(Optional) Numeric vector. A vector of fleet effort on a log-scale.}

\item{maxRetry}{(Optional) Integer. How many times to re-run optimisation if
identification of choke-stocks is unsuccessful? Default is 1.}

\item{useEffortAsInit}{(Optional) Boolean. Should the fleet effort from the
previous year be used as initial values. Default is
\code{FALSE}}

\item{useGlobalAsInit}{(Optional) Boolean. Should the fleet effort estimated
from initial Global Optimisation be used as initial
values for subsequent Choke Optimisation.}

\item{useTMB}{(Optional) Boolean. Should TMB be used to estimate fleet effort?
Default is \code{TRUE}. Highly recommended that the default is
used.}

\item{correctResid}{(Optional) Boolean. Should fleet efforts be scaled down
to bring residual overquota catch to zero if residual
over-quota catch following optimisation exceeds 0.001?
Defaults to \code{FALSE}}

\item{parallel}{(Optional) Boolean. Should function be run in parallel?}

\item{verbose}{(Optional) Boolean. Should messages be printed to console?
Defaults to \code{FALSE}}

\item{method}{(Optional) Character. Determines the optimisation algorithm that is used.
Values may be 'nlminb' or 'Nelder-Mead'. Default is 'nlminb'.
HIGHLY recommended that the default is used.}
}
\value{
A list containing outputs for the final optimisation.
}
\description{
This function calculates the effort for each fleet required to catch their
most-limiting or least-limiting quota, as specified using the \code{effortType}
argument.
}
\details{
Outputs are a list of the form:

\itemize{
\item A list of iterations
\itemize{
\item \code{par}: A vector of log-scale effort per fleet
\item \code{objective}: The final optimisation objective function value.
\item \code{convergence}: The final optimisation convergence code value.
0 indicates successful convergence.
\item \code{iterations}: The final optimisation number of iterations.
\item \code{message}: The final optimisation convergence message.
}
}
}
