% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_effortdyn_FLBiol_v2.R
\name{FLBiols2List}
\alias{FLBiols2List}
\title{Convert object into simple nested list structure}
\usage{
FLBiols2List(
  om,
  year,
  advice,
  useCpp = TRUE,
  avgE_nyear = 1,
  process_residuals = NULL
)
}
\arguments{
\item{om}{A list of \code{FLBiols} and \code{FLFisheries} containing the relevant
stock and fleet information in the Operating Model.}

\item{year}{Integer. Year in which advice is implemented.}

\item{advice}{A list of the total allowable catch or landings advice for each
stock.}

\item{useCpp}{(Optional) Logical. Should data be processed using faster C++ code?
Defaults to \code{TRUE}.}

\item{avgE_nyear}{Integer. Number of reference years for the calculation of average
fleet effort.}

\item{process_residuals}{Defaults to \code{NULL}.}
}
\description{
This function converts the linked \code{FLBiols} and \code{FLFisheries}
objects into a set of nested lists containing the data required by
\code{effortBaranov} to optimise fleet efforts to fully catch their first quota.
}
\details{
Outputs take the following structure:

\itemize{
\item \code{iterations}: a named list
\itemize{
\item \code{n}: a named list
\itemize{
\item vector of numbers-at-age for ith stock
}
\item \code{m}: a named list
\itemize{
\item vector of natural mortality-at-age for ith stock
}
\item \code{landwt}: a named list
\itemize{
\item matrix of landed individual mean weight-at-age for ith stock (age x fleet)
}
\item \code{discwt}: a named list
\itemize{
\item matrix of discarded individual mean weight-at-age for ith stock (age x fleet)
}
\item \code{landfrac}: a named list
\itemize{
\item matrix of proportion of size-selective-retained catch-at-age for
ith stock (age x fleet)
}
\item \code{catchsel}: a named list
\itemize{
\item matrix of catch selectivity-at-age for ith stock (age x fleet)
}
\item \code{catchq}: a matrix of catchability (stock x fleet)
\item \code{quotashare}: a matrix of proportional quota-share (stock x fleet)
}
}
}
\examples{
## load example data
data("mixedfishery_MixME_input")

## run MixME simulation
out <- FLBiols2List(om = mixedfishery_MixME_input$om,
                    year = 2020,
                    advice = list(cod = 1000, had = 1000))
}
