% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_fwd.R
\name{fwdMixME}
\alias{fwdMixME}
\title{Forward projection of stocks - including process error}
\usage{
fwdMixME(
  om,
  ctrl,
  args,
  tracking,
  sr_residuals_mult = TRUE,
  effort_max = 100,
  proc_res = NULL,
  ...
)
}
\arguments{
\item{om}{A list of \code{FLBiols} and \code{FLFisheries} containing the relevant
stock and fleet information in the Operating Model.}

\item{ctrl}{Control object.}

\item{args}{list of additional arguments}

\item{tracking}{a named list of tracking objects to monitor emergent dynamic
properties}

\item{sr_residuals_mult}{(Optional) Logical. Are stock recruitment residuals
multiplicative? Defaults to \code{TRUE}.}

\item{effort_max}{(Optional) Numeric value indicating the maximum allowed
fishing effort for any fleet. Defaults to 100 and
cannot be \code{NULL}.}

\item{proc_res}{(Optional) Character. Where is process error noise stored?
If \code{NULL}, no process error is applied to stock numbers.
Defaults to \code{NULL}.}
}
\value{
A list containing the \code{FLBiols} and \code{FLFisheries} objects
with projected stock numbers, fleet efforts, and fleet-stock landings
and discards numbers.
}
\description{
This function takes the output of the implementation error model and
projects the stocks and fleets forward one time step.
}
