% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_simulation_makeTracking.R
\name{makeTracking}
\alias{makeTracking}
\title{Generate a tracker to store emergent simulation properties}
\usage{
makeTracking(om, projyrs, addmetrics = NULL)
}
\description{
This function generates a tracking object to store diagnostics and emergent
properties generated during the MSE simulation.
}
\details{
By default, containers for the following properies are generated. Additional
containers are generated for user-supplied...
}
