% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{match.c}
\alias{match.c}
\title{Relabel the Clusters of Membership Probability}
\usage{
match.c(X, Y)
}
\arguments{
\item{X}{A matrix of size n*K consisting of membership probabilities}

\item{Y}{A matrix of size n*K consisting of membership probabilities}
}
\value{
Reordered Y by its columns, same cluster order as \code{X}
}
\description{
return matched Y corresponding to X
}
