% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_multiSAM2FLIndex.R
\name{multiSAM2FLIndex}
\alias{multiSAM2FLIndex}
\title{Extract survey data from a SAM fitted object to an \code{FLIndices} object}
\usage{
multiSAM2FLIndex(
  SAMfit,
  yearRange = NULL,
  uncertainty = FALSE,
  niter = 1000,
  samVariates = NULL,
  seed = NULL
)
}
\arguments{
\item{yearRange}{}
}
\value{
an \code{FLIndices} object
}
\description{
Function takes a fitted SAM object as input and returns an \code{FLIndices}
object containing the observed survey index values and the fitted
stock catchability.
}
