% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_multiSAM2FLR.R
\name{multiSAM2FLR}
\alias{multiSAM2FLR}
\alias{multiSAM2FLR,sam-method}
\alias{multiSAM2FLR,sam_list-method}
\title{Convert SAM fitted object into \code{FLStock} and either
\code{FLFisheries} or \code{FLFleets} objects}
\usage{
multiSAM2FLR(
  SAMfit,
  stkname = NULL,
  useSAMcatch = TRUE,
  uncertainty = FALSE,
  niter = 1000,
  seed = NULL,
  ...
)

\S4method{multiSAM2FLR}{sam}(
  SAMfit,
  stkname = NULL,
  useSAMcatch = TRUE,
  uncertainty = FALSE,
  niter = 1000,
  seed = NULL
)

\S4method{multiSAM2FLR}{sam_list}(
  SAMfit,
  stkname = NULL,
  useSAMcatch = TRUE,
  uncertainty = FALSE,
  niter = 1000,
  seed = NULL,
  ...
)
}
\arguments{
\item{SAMfit}{A SAM fitted stock assessment model object or
a named list of SAM fitted stock assessment model objects}

\item{stkname}{Stock name to be used in  \code{FLFishery} or \code{FLFleet}
object. Ignored if object of class \code{sam_list} is supplied.}

\item{useSAMcatch}{Optional argument. If \code{TRUE}, the fitted catches estimated
by SAM are used. Otherwise, the observed catches are used.
Defaults to \code{TRUE}.}

\item{uncertainty}{Optional argument. If \code{TRUE}, the iteration dimension
of the returned objects contains replicates (uncertainty)
sampled from the variance-covariance matrix of the fitted
SAM object. Defaults to \code{FALSE}}

\item{niter}{Optional integer. The number of replicates sampled if
\code{uncertainty} is \code{TRUE}. Default is 1000 replicates.}

\item{seed}{Optional integer. This is a random number seed to generate
reproducible outputs. Defaults to \code{NULL}.}

\item{fltClass}{Class structure to be used for fleets, defaults to
\code{FLFishery}. Either \code{FLFishery} or \code{FLFleet}}
}
\value{
A list containing an \code{FLStock} and \code{FLFleets} objects
}
\description{
This function takes a SAM fitted stock assessment object as input and
returns a list of \code{FLStocks} and either \code{FLFisheries}
or \code{FLFleets} objects.
}
\details{
Returned \code{FLStock} objects contain data for:
\itemize{
\item stock numbers-at-age (\code{stock.n})
\item catch numbers-at-age (\code{catch.n})
\item landings numbers-at-age (\code{landings.n})
\item discards numbers-at-age (\code{discards.n})
\item proportion mature-at-age (\code{mat})
\item natural mortality-at-age (\code{m})
\item fishing mortality-at-age (\code{harvest})
\item stock mean weight-at-age (\code{stock.wt})
\item catch mean weight-at-age (\code{catch.wt})
\item landings mean weight-at-age (\code{landings.wt})
\item discards mean weight-at-age (\code{discards.wt})
}

Catch numbers and fishing mortality-at-age are summed across commercial fleets,
whereas catch, landings and discards mean weight-at-age are averages weighted
by the catch number proportions by commercial fleets. Landings and discards
numbers-at-age are calculated using the catch proportion weighted mean
landings fraction.

The function also updates the Fbar range for the stock from the fitted
SAM object.

Returned \code{FLFleet} objects contain data for:
\itemize{
\item partial fishing mortality-at-age (\code{harvest})
\item catch numbers-at-age (\code{catch.n})
\item landings numbers-at-age (\code{landings.n})
\item discards numbers-at-age (\code{discards.n})
\item catch mean weight-at-age (\code{catch.wt})
\item landings mean weight-at-age (\code{landings.wt})
\item discards mean weight-at-age (\code{discards.wt})
}
}
\section{Warning}{

This function requires \code{FLCore} and either \code{FLFishery} or
\code{FLFleet} to operate.
}

