% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_oem.R
\name{oemMixME}
\alias{oemMixME}
\title{Implementation of stock observation methods}
\usage{
oemMixME(
  om,
  deviances,
  observations,
  args,
  tracking,
  catch_timing = NULL,
  idx_timing = NULL,
  use_stk_oem = FALSE,
  use_catch_residuals = FALSE,
  use_idx_residuals = FALSE
)
}
\arguments{
\item{om}{An operating model}

\item{deviances}{...}

\item{observations}{...}

\item{args}{List of additional arguments}

\item{tracking}{A named list of tracking objects to monitor emergent dynamic
properties}

\item{catch_timing}{A named list of catch timing relative to assessment year}

\item{idx_timing}{A named list of survey index timing relative to assessment year}

\item{use_stk_oem}{Logical vector. Length \code{n} stocks. Should a supplied
Observation Error Model structure be used?}

\item{use_catch_residuals}{Logical vector. Length \code{n} stocks. Should
catch residuals be used to generate uncertainty?}

\item{use_idx_residuals}{Logical vector. Length \code{n} stocks. Should
survey index residuals be used to generate uncertainty?}
}
\value{
A named list of stock observations, survey indices and updated tracking
object
}
\description{
Function to generate stock observations (fleet catches and survey indices)
with (optional) uncertainty.
}
