% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plotting_timeseries.R
\name{plot_timeseries_MixME}
\alias{plot_timeseries_MixME}
\title{Plot time-series of Operating Model properties}
\usage{
plot_timeseries_MixME(
  object,
  quantity,
  minyr = NULL,
  maxyr = NULL,
  stknames = NULL,
  fltnames = NULL,
  trajectories = NULL,
  quantiles = c(0.05, 0.25, 0.75, 0.95),
  addRefpts = TRUE,
  keepFailedIters = TRUE
)
}
\description{
Function takes an Operating model as input and plots a time-series of
a specified property for one or more stocks or fleets. Annual distributions
are summarised to the median and up to two user-specified quantile intervals.
}
