% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plotting_trajectories.R
\name{plot_trajectories_MixME}
\alias{plot_trajectories_MixME}
\title{Plot individual trajectories of Operating Model properties}
\usage{
plot_trajectories_MixME(
  object,
  quantity,
  minyr = NULL,
  maxyr = NULL,
  stknames = NULL,
  fltnames = NULL,
  addRefpts = TRUE,
  failedIters = "highlight"
)
}
\description{
Function takes an Operating model as input and plots a time-series of
a specified property for one or more stocks. Individual trajectories are
plotted and failed trajectories are optionally highlighted or removed
}
