% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_conditioning_stfMixME.R
\name{stfMixME}
\alias{stfMixME}
\title{Function to extend a MixME Operating Model to future years for forecasting}
\usage{
stfMixME(
  om,
  nyears = 3,
  wts.nyears = 3,
  sel.nyears = 3,
  qs.nyears = 3,
  verbose = TRUE
)
}
\description{
Function to extend a MixME Operating Model to future years for forecasting
}
